Fix root path for explore usage.
